Fix BSPSceneManager, was broken by a recent signature change last week but missed ANSI compliance - remove anonymous structs from Vector2/3/4, Quaternion, ColourValue. 'val' is gone, use ptr() for pointer access, [] for array access Warning cleaning - OgreMain now builds with no warnings on VC8 /W3 Small tidy ups

+++ OgreMain/include/OgreImageCodec.h
@@ -53,10 +53,10 @@ namespace Ogre {
 				num_mipmaps(0), flags(0), format(PF_UNKNOWN)
 			{
 			}
-            uint height;
-            uint width;
-			uint depth;
-            uint size;
+            size_t height;
+            size_t width;
+			size_t depth;
+            size_t size;
             
             ushort num_mipmaps;
             uint flags;
