+++ OgreMain/include/OgreTexture.h
@@ -133,40 +133,40 @@ namespace Ogre {
 
         /** Returns the height of the texture.
         */
-        virtual unsigned int getHeight(void) const { return mHeight; }
+        virtual size_t getHeight(void) const { return mHeight; }
 
         /** Returns the width of the texture.
         */
-        virtual unsigned int getWidth(void) const { return mWidth; }
+        virtual size_t getWidth(void) const { return mWidth; }
 
         /** Returns the depth of the texture (only applicable for 3D textures).
         */
-        virtual unsigned int getDepth(void) const { return mDepth; }
+        virtual size_t getDepth(void) const { return mDepth; }
 
         /** Returns the height of the original input texture (may differ due to hardware requirements).
         */
-        virtual unsigned int getSrcHeight(void) const { return mSrcHeight; }
+        virtual size_t getSrcHeight(void) const { return mSrcHeight; }
 
         /** Returns the width of the original input texture (may differ due to hardware requirements).
         */
-        virtual unsigned int getSrcWidth(void) const { return mSrcWidth; }
+        virtual size_t getSrcWidth(void) const { return mSrcWidth; }
 
         /** Returns the original depth of the input texture (only applicable for 3D textures).
         */
-        virtual unsigned int getSrcDepth(void) const { return mSrcDepth; }
+        virtual size_t getSrcDepth(void) const { return mSrcDepth; }
 
         /** Set the height of the texture; can only do this before load();
         */
-        virtual void setHeight(unsigned int h) { mHeight = mSrcHeight = h; }
+        virtual void setHeight(size_t h) { mHeight = mSrcHeight = h; }
 
         /** Set the width of the texture; can only do this before load();
         */
-        virtual void setWidth(unsigned int w) { mWidth = mSrcWidth = w; }
+        virtual void setWidth(size_t w) { mWidth = mSrcWidth = w; }
 
         /** Set the depth of the texture (only applicable for 3D textures);
             ; can only do this before load();
         */
-        virtual void setDepth(unsigned int d)  { mDepth = mSrcDepth = d; }
+        virtual void setDepth(size_t d)  { mDepth = mSrcDepth = d; }
 
         /** Returns the TextureUsage indentifier for this Texture
         */
@@ -313,9 +313,9 @@ namespace Ogre {
 		virtual HardwarePixelBufferSharedPtr getBuffer(size_t face=0, size_t mipmap=0) = 0;
 
     protected:
-        unsigned long mHeight;
-        unsigned long mWidth;
-        unsigned long mDepth;
+        size_t mHeight;
+        size_t mWidth;
+        size_t mDepth;
 
         size_t mNumRequestedMipmaps;
 		size_t mNumMipmaps;
@@ -327,7 +327,7 @@ namespace Ogre {
         int mUsage; // Bit field, so this can't be TextureUsage
 
         PixelFormat mSrcFormat;
-        unsigned long mSrcWidth, mSrcHeight, mSrcDepth;
+        size_t mSrcWidth, mSrcHeight, mSrcDepth;
 
         PixelFormat mDesiredFormat;
         unsigned short mDesiredIntegerBitDepth;
